added blockList UI
