--- app/src/features/settings/components/GlobalSettings/components/BlockList/blocklist.scss
@@ -1,4 +1,0 @@
-.blocklist-body {
-  background: var(--requestly-color-surface-0);
-  border-radius: 8px;
-}
