+++ app/src/features/settings/components/GlobalSettings/components/BlockListSettings/blocklist.scss
@@ -1,0 +1,28 @@
+.blocklist-body {
+  background: var(--requestly-color-surface-0);
+  border-radius: 8px;
+  padding: 16px;
+  .header-row {
+    padding-bottom: 8px;
+
+    .rq-input {
+      height: 32px;
+    }
+  }
+
+  .blocklist-subheading {
+    color: var(--requestly-color-text-subtle);
+    font-size: var(--requestly-font-size-xs);
+    padding: 8px 0;
+  }
+
+  .blocklist-domains-container {
+    padding: 8px;
+    line-height: var(--requestly-font-line-height-lg);
+    border-bottom: 1px solid var(--requestly-color-white-t-10);
+
+    .blocklist-close-icon {
+      cursor: pointer;
+    }
+  }
+}
